Added model endpoint pallete IDs and CSS attributes
Necessary to pick up latest additions to the API which picks up the `palette_ids` and `css` attributes. This also removes the `published_at` date/time since it's used and provides little value since Core never records the exact publish date of when the model/device was released.

Milestone: minor