Added IP address contract
Necessary to validate `ips` endpoint payloads.

RuboCop has been updated to ignore the use of `ipv4` and `ipv6` since this is a design flaw of the original API design since `version_4` or `version_6` would have been better. This'll soon be corrected (transformed) when the corresponding model is added.

Milestone: minor